Huge functionality rewrite
Rewrite half of the entire launcher. It still operates under a
somewhat similar basis as its previous version (slide
somewhere to get a list of apps), but now it's functionally
better and is actually usable as a launcher.

Changes that were made in this commit:
* Replace our dirty toolbar hack with a sliding panel.
* Make the launcher show the actual wallpaper instead of fitting a
wallpaper inside an ImageView. This fixes crazy cropping we
see in the previous implementation.
* Add 'Uninstall' and 'App info' to app list to allow for
quick(er) operations.
* Rewrite a good chunk of features.
* Remove customisation options that doesn't work with the new
panel.
* Clean up a whole bunch of dead code.
* Bug fixes and speedy performance improvements.

Finally an actual change after decades of hibernation.