Guard search bar launch against empty appList
When searching, it is possible for the filter or appList to have an
empty object count since search query can hit no target. However,
because the launch here is unguarded, the search key will still try to
launch the first object in appList, leading to a crash.

Fix this by checking for appList emptiness before launching.