Set the duration for search bar fade animation
I noticed that in some devices, when the bar is pulled down, the
animation doesn't seem to finish quickly enough, resulting the bar
popping up from the aether (thus, translating to a bad transition
animation).

Fix this by setting a fix duration for the bar pull and return
animation. Its return animation should be longer, as it needs to travel
the full screen distance, whereas its pull animation should be shorter,
since it will be the first element the user will interact with.