Trim search bar text query
It's possible (even if without a reason) that a user enters a spam of
space to the search bar. This is fine, but it will also prompt the user
if they want to search for said spam of space online — resulting in a
bad UX experience.

To overcome this, we simply need to trim the text, as it will remove
leading and trailing space, effectively fixing our problem here.