Refactor hidden app management
Split the hidden app management screen into an activity because it's
clear that maintaining it as a PreferenceFragment is completely wrong
and wouldn't work. This is technically still a hack, but it's a much
better one than before. Considering that everything works as expected,
I'll keep this until a better solution is found.