Redo AppAdapter refactoring
The previous commit, 3ae7c9c0ff4944e18c0197ef0371f1c910b133e3,
completely breaks searching because the filter relies on interfacing
directly with the adapter's list — that very same list that gets cleared
whenever a search query is entered. This causes the search result to
replace the app list.

Knowing that, it's better to revert to the previous revision of the
filter method. Our issue was only that the filter wasn't following the
changes made to the list, so instead of refactoring the entire thing,
simply have methods to force a refresh of the filter.

With this, app hiding works fine and search mechanism is undisrupted.