Automatically update app list when a hidden app is restored
In loadApps, it's nearly impossible to do a complete refresh without
notifyDataSetChanged, and trying to call it from preference change
listener crashes the launcher. So, along with the listener change, also
introduce a way to forcefully refresh the app list without making it
crash — thus allowing a faster restoration of hidden apps.