Fix crashes and odd behaviour with app hiding
If the user attempts to hide an app that is at the very end of the app
list, they will experience a crash regarding incorrect position of the
ViewHolder. I thought at first, this can be solved by doing:

appList.remove(position - 1);

However, while this does work, it causes an odd behaviour where the
launcher will also remove the app above the app user has selected.

A workaround to this is to use another absolute index we have: package
name. Because there is no two package name that can be the same, it can
be used as a reference for list removal. There may be some performance
impact to this, as opposed to using list, but it shouldn't be
negligible.