Workaround for Oreo broadcast limitation
A more proper way to fix the limitation is to poll for changes from
PackageManager. The polling would be controlled by JobScheduler, thus
making the app less toxic to Doze and Oreo's background management.

However, because JobScheduler (and to an extent, JobService) can be
not-so-reliable, let's instead reuse PackageChangesReceiver by
registering it in MainActivity. We're killing two problems with one
throw.