Unbreak wallpaper shade on v21+
Shade doesn't go all the way through in Lollipop due to the difference
in statusbar style compared to KitKat's transparent statusbar — fix this
by overriding several flags and manually managing insets. This does
require a workaround for the app panel as now the statusbar is
completely transparent. I have made a padding to handle it, but I feel
that it can be unfit to look at for some.