Implement alternate padding mode for app panel
This looks better in my opinion, but make it optional because there
isn't any reason for it not to be an option.

This commit also splits shade from touchReceiver. While the animation
touchReceiver gets while it is being shaded is neat, making it one with
touchReceiver also hampers  further customisation. Separate them both
because that's how things should have been from the start.