Fix erratic search bar in alternate padding mode
When search bar is dragged up or pulled down, there will be a brief
period where it breaks and begin to pads itself, causing the corners of
the search bar to be cut or be shifted to its edges.

Unsurprisingly, this is rather simple to fix. Swapping out padding for
margin does the job, so let's do that.

* Also remove useCompatPadding; it makes the search bar looks tiny in
v21+.