Unbreak wallpaper shade again
Wallpaper shade doesn't expand all the way through the navigation bar.
Unlike the statusbar, however, we don't need a huge hack to fix this —
simply set the navigation bar colour to a matching shade of black. The
shade blends perfectly with the wallpaper shade, so there shouldn't be
any visible visual disruption.