Temporary fix for keyboard killing search bar
Reproducing the issue is fairly easy:
* Pull down the app panel
* During the animation, quickly tap on the app list
* Go back to the launcher; it should be frozen and the search bar should
stuck in whatever state it was in
* Click on the search bar, only to be met with it disappearing and the
user unable to pull the app panel away

The issue seems to be caused by the keyboard or animation forcefully
locking the panel. This is a temporary workaround for the issue and
further digging is needed to confirm the cause.