Replace fuzzy search implementation
The previous implementation is, as noted, not too great for app
searching. Levenshtein distance is good for matching strings, but it
gets too many false positives, even when weighed.

In my opinion, KISS (to an extent, Sublime Text) has a good
implementation for fuzzy search. They have moved on from this particular
implementation, but for Hubby, this one is already an improvement.