Completely remove hidden_apps listener
The only reason why we were listening to it in the first place was to
make sure restored apps are being properly refreshed. However, because
of the new flag, it is not needed to listen for it anymore.

Added to that, listening for it here is useless. App hiding mechanism
(removal from list, etc) is already handled in recycler click, so there
is even less point for its existence.

This should also fix the odd app hiding issue plaguing Jelly Bean (and
possibly others).