Workaround for broken app installation filter
It is possible that the user would have an app installing in the
background while they are typing a query in search. When this happens
and the app is installed while the query is being typed, the app list
will completely lose all its content until onResume is called (and a
refresh is done again).

This used to be a non-issue when loadApps() is used, but since the
advent of loadSingleApp, the issue surfaces quite often.

While we can just call loadApps again, it's bad for the UX as it blocks
the entirety of the app interaction — causing it to appear frozen. So
instead, let's recreate the activity; it is cheaper and less obtrusive
for the UX.