Set top margin for favourites bar
My first idea for the favourites bar was for it to extend across the
screen seamlessly. However, due to limitations imposed by Google when it
comes to drawing underneath statusbar, the idea seems to be extremely
unfeasible.

So, with that in mind, we can scale down the panel so it extends less to
the statusbar. It still looks somewhat okay-ish, after all.

Thanks Google for ruining a maiden dream.