Fix filter getting stuck when an app is hidden
When a search is performed and the user chooses to hide an app while the
query is still enabled, the app list will be stuck to displaying only
the result of the query. This is the same issue as before: filter is
completely unreliable and unresponsive at times.

This should solve parts of the issue, but there should still be an
attempt to improve the filter we already have.