Early implementation for backup and restore
Phew, this is quite the hacky solution. The implementation is very
simple in code: fetch preferences, write them to an XML, and when
restored we set their keys back to the preferences. This shows to be
reliable enough and it doesn't overcomplicate everything.

For now, this suffices and UX problems coming from it will be sorted out
soon.