Separate dimensions for favourites panel
When the panel lays at the bottom, its size looks so small and the touch
area is wholly unoptimised in smaller screens.

To fix this, have a separate size for vertical and horizontal panel. We
can set the
horizontal panel to be bigger than its vertical counterpart to increase
touch area.