Disable extract UI in all EditText
It's very intrusive when the landscape orientation is used. Not only
that it breaks the majority of the user experience flow, as it switches
over to a wholly different screen when EditText is focused, it also
isn't just all that great to have for the launcher.

Kill it and now we're all happy.