Disable recycling for MenuPreference
During testing in Oreo, I notice that preferences using MenuPreference
were inconsistently unstable. It would crash due to PopupMenuHelper not
being able to get any anchor for the PopupMenu, and this is definitely
an API-isolated behaviour as anything below O works fine.

Apparently, the preferences recycling behaviour in Oreo caused IDs of
the preferences to be broken or straight up missing/null.

Disable this new behaviour as I don't want this to become a random crash
fest.