Stop keyboard from pushing app list up
This is especially annoying when the user is trying to reach for an app
at the very bottom of the list. Setting windowSoftInputMode to adjustPan
is one of the workarounds, but adjustPan breaks the view entirely.

A proper fix to this issue is to just listen for layout changes, and
forcefully scroll down when the keyboard pushes the view up.