Use slidingHome margin for statusbar workaround
Setting only the app list's padding will make the search bar go below
the statusbar when pulled. This may be the expected behaviour, sure, but
it breaks when the statusbar isn't transparent (i.e when battery saver
mode is enabled) — when this case is met, the search bar will have half
of its view cut by the statusbar until it is pulled down.

Fix this by having slidingHome set its margin (not padding) the same as
the statusbar's height. With this, everything will render below the
statusbar but not underneath it. This change also wouldn't interfere
with the wallpaper shade, as it is on a completely separate view parent.