Apply theme changes immediately
Going to the desktop and going back to settings seems like a huge hassle
from a user-standpoint. However, if we simply recreate the activity to
immediately toggle for theme changes, it'll cause a bizarre and abrupt
transiton.

The fix is rather simple: set enter and exit animation and let Android
call the activity again to play the transition animation. With this, we
are able to get a smooth theme transition that looks at least visually
interesting.