Don't kill spaces when fuzzy searching
Removing spaces will make us lose some relevant search results.

An example of this is 'Better VPN', which has two potential query that
can be done to find it: 'better' and 'vpn'. If the user were to search
for 'VPN', they won't find any result, because the 'Better VPN' is being
processed as 'bettervpn' — not 'better + vpn'.

This problem is exaggerated by how the fuzzy search works: first-letter
match has a higher score, which means that 'VPN', which is placed at the
very end in 'bettervpn' will not at all yield any result due to its low
score.