Rework favourites panel layout
There were lots of issues regarding the favourites panel. Most
importantly is just how plain bad it looks and how hard for it to be
maintained every time there is a new layout change.

Another persistent problem was its lack of visual cue. We may be able to
fix this, but with how everything is and how redundant it feels with its
swiping, I think it's better to part ways with its old layout.

The new layout integrates it with the search bar. It should appear on
top of search bar, which makes it more visible and also easier to
utilise.