Use static Y value for scroll detection
Using delta is only good for scrolling up, since users can scroll up or
down minutely/slowly, which causes the animation to stutter as it is
trying to switch between being hidden to being visible rapidly.

Setting these to a static value should counteract these jitters. Now the
panel will react to a long scroll/a fling and not a small, incremental
swipes.