Make search snackbar less hyperactive
To start with, the snackbar has several problems that needed addressing:

1.) It's not automatically dismissed if search query is emptied out.
2.) If the user rapidly enters their query, there is a possibility that
it'll dismiss itself; this is because snackbars will automatically
dismiss any other existing snackbar if they exist.
3.) It's just a bit too jumpy overall.

This commit address and fixes all the above. The snackbar should now
only change its query instead of dismissing and recalling itself; it
makes for a better experience and it won't try to kill itself whenever
the user is making a lightning-fast query.