Fix wrongly dismissed snackbar
When the user taps a suggestion strip in their keyboard, code-wise, the
search bar momentarily goes empty; this dismisses the snackbar, which
shouldn't happen.

A quick way to fix this is to also check for the query's (s) length. The
query is always updated when a new text is entered, and thus, it should
prevent the snackbar from dismissing early.