Disable scroll hide/show animation for favourites panel
There is an issue where overscrolling the app list will cause the
animation to jump all over the place. Until we can find a fix to it,
let's temporarily disable it for now.

Animations will still show in some actions, such as when the panel is
returning or when it is hidden due to last item being unpinned. It's
just that we won't have any when scrolling.