Create toolbar when there isn't any
The preference library we're using doesn't quite manage fragments that
doesn't inherit its classes. As such, for devices running KitKat and
below, toolbar will be nonexistent in these fragments. Lollipop gets a
free pass because the library delegates everything to the support
library, thus allowing  for the toolbar to be automagically generated.