Fix inconsistent toolbar text colour
In some devices running TouchWiz versions earlier than KitKat will
pretty much tint the text colour of toolbar title and action menu black.
This is not the intended colour, so it better know what it is for the
sake of uniformity and consistency (and because black doesn't look nice
here).