Rely on favourites panel to listen for layout resize
We had an old issue where the keyboard would push the search bar up, but
the app list wouldn't scroll down — causing the app list to be covered
by the search bar. Our fix to this was to listen for layout change in
the app list.

However, this has proven itself to be extremely unreliable, now that
favourites panel is in the same view group as the app list. When the
keyboard is dismissed in an unexpected state (such as when the list has
been scrolled), it will reset the scrolling position *and* cause the
favourites panel to appear in where the search bar was before the
keyboard was dismissed.

So, instead of listening for layout changes in the app list, we now
listen in favourites panel instead. This is not only more reliable, as
the favourites panel has a static size, it also fixes all the problem
listed above.