Fix app list pushing when favourites panel is gone
Because we rely on the favourites panel to listen for layout changes,
when it is gone (not invisible), it will stop participating in listening
for changes — as the system exempts it from drawing. This will make the
app list stay unpushed when the keyboard is showing, thus, reverting us
back to our old issue.

To fix this, as a temporary measure, we let the app list listen for
layout changes like before — but, we only do so conditionally.