Improve app list scroll detection
In our previous implementation, animation plays almost inconsistently;
the hide animation never plays, and the show animation plays only when
there hasn't been a rapid fling done.

In this commit, I devise this small copy of HidingScrollListener, which
listens to only scroll up as well as list end. With this scroll
listener, show animation would play but hide animation still never shows
itself.

The hide animation should be fixed at a future commit, but for now, this
is enough improvement as it is.