Launch apps depending on scroll position
If there is nowhere to scroll, up or down, this means the search result
has been narrowed down and the best match is usually presented down at
the bottom.  If we were to use search to launch the app, it's best to
consider this positioning; the user shouldn't launch apps that they
don't see (i.e the one at the top), they should launch either the top or
the bottom — depending on which one is visible.