Fix PackageReceiver register/unregister log spam
Half of the time, logcat is completely flooded with errors such as:

- E/ActivityThread: Activity mono.hg.MainActivity has leaked
IntentReceiver.
- java.lang.IllegalArgumentException: Receiver not registered.

These are all receiver-related issue that we have been having for quite
a long time now, and means of workaround has been made for it. However,
the issue would still pops up because of the irregular cycle of the
receiver.

This commit is meant to hack a fix to the issue. Instead of relying the
activity to unregister the receiver, we let the receiver do that; after
all, once we send the signal that an app has been installed/uninstalled,
we'll be passing onCreate anyway — and that will re-register the
receiver again, thus the receiver lifecycle is better maintained.

This is not how it should have been done, but as of now, it silences all
of these errors without any huge side effect — so we're keeping it for
the time being.