Switch PinnedApps to FlexibleAdapter
Pinned apps (or rather, the object, PinnedApp) is similar enough to the
normal app list (AppDetail) — we can extend from it and kill off parts
we don't need to shorten our LoC. With this, we can refactor some longer
and more tedious parts of our code such as loadSingleApp.

I'll be keeping loadSingleApp for now — this is because in the later
future I may end up stop (ab)using recreate(), so if that ever happens,
loadSingleApp will be useful again.