Implement basic icon shading for adaptive icons
This commit tries to fix two different problem with one shot, and those
problems are:

* White icons are blending with the app list background with light
theme.
* The lack of shading for adaptive icons; this follows AOSP Launcher3
which has adaptive icons shaded.

For the conflicting background colour, I have restyled the light theme.
It should now look a tad bit better and white-on-white shouldn't happen
anymore.

As for the shading, while I can try implementing much more complex
methods — I feel that a simple drawable/bitmap shading is enough until a
need arises for me to refactor these. This works well enough and the
performance drop is largely unseen (but it should still be held with
caution).