Properly unregister/register receiver
Before we had onCreate and onResume create them and then we let the
receiver itself handle the unregistering. This worked for a short while,
and there isn't any problem with it.

However, upon further inspection, it seems clear that the leak is still
happening. We don't want that, so we resort to our worst case scenario:
the receiver is unregistered regardless while we ignore the
IllegalArgumentException being thrown by the nonexistent receiver.