Fix pinned apps order persistence
We shouldn't expect order from StringSet; this is ingrained in how Sets
work in Java: they are unordered lists.

I can make a sorting system, sure, but that is too much for a simple
list. Additionally, it doesn't warrant further maintainability as we are
going against the design of Set itself.

The common solution seems to be to replace StringSet altogether. Using
concatenated string, we can then take a delimiter and split the strings
accordingly — doing an array whenever we need to fetch its content.

As far as performance and usability foes, this is a fine solution. It
should also allow us to have a chance at item order swapping, but we'll
have to see about that.