Don't make search snackbar rely on favourites panel
When favourites panel is disabled, snackbar will disappear along with
it, as the container is the one getting killed — this container is the
one hosting snackbar, thus the root of the issue.

As a workaround, wrap the favourites panel in its own FrameLayout. With
this, it can freely be hidden without affecting the search snackbar.