Fix EditText hint colour
The previously-used 'secondary_text_dark' is a colour state list — which
means it responds to events such as touches and taps. While that is
perfectly fine to have, the colour used to indicate those events clashes
with our EditText background, causing bad UX.

Change this to a static colour for now, as that should work perfectly
fine. This should also cure someone's nitpick as well.