Don't restore search bar when panel is invisible
When panel is invisible, the assumption is that the search bar would be
invisible too: but this is not the case. The search bar can be visible
here, showing on the top even though the panel is not pulled down.

Check for panel visibility here. If it's invisible then don't restore
the state — leave the search bar invisible as well.