Only filter apps when they have been loaded
Have a dummy boolean in the adapter that notifies the activity when the
adapter has loaded its data. When this flag has been set to true (as in
onPostExecute), the filter can start working again.

This is a fix to the issue where it is possible for the user to filter
out an empty adapter when the app list is still loading. We don't want
to stop the user from using the search bar (which is also a feasible
workaround), so instead, we simply prevent filtering from ever happening
when the app list is still loading.