Convert over to ComponentName
Package name works for the most part, but ComponentName allows greater
range of precision. One issue this fixes is when the user tries to hide
an app with multiple launch intent, such as Google. When using only the
package name, the launcher will hide all launch intent from the app,
whereas if we use ComponentName, the launcher will only hide that
specific class.

This commit is largely experimental and more can be tested to make sure
the app still functions as expected.