Set FLAG_ACTIVITY_NEW_TASK before launching an app
It seems that when using Intent.makeMainActivity, some app outright
refuse to launch. One good example to this is Tachiyomi, who, despite
not throwing any exception or error, will not display any of its
activity at all when launched.

A workaround I had in mind was to reuse the makeRestartActivityTask, but
that breaks app transition as a whole — causing apps to momentarily show
a blank white screen when launched.

So it seems that we still need this flag for the time being. This fixes
the issue at least and it doesn't show any unexpected behaviours.