Unbreak widget events
Widget events was disabled by overlaying touchReceiver on top of the
widget host. This was done because the whole launcher is rendered
useless when a widget takes the entire space of the screen — the user
would not be able to remove the widget either as that require calling
for the context menu, which the user can't do since the touchReceiver is
obscured by the widget.

AOSP Launcher actually has a fix for this. Using their custom
LauncherAppWidget classes, the widgets are able to receive long clicks.
With this, the context menu can be called regardless of whether or not
touchReceiver is visible.

Swiping (may) still be broken, so calling the app panel down is still
out of the question. There needs to be more fixes for this in the near
future.