Make widget visibility toggle less jarring
If we immediately hide the widget when the panel is dragging then the
widget will simply pop back and in out of existence. To lessen the
horrible eye sore, let's only hide the widget once we know that the
panel has been fully expanded — this makes the widget animate along the
panel as well.