Fix leaking and respawning receivers
In Android P, the launcher was repeatedly crashing due to the respawning
receiver which the system does not tolerate nicely. This issue has been
plaguing the launcher anyhow, and it seems that Android P is telling us
to stop ignoring the issue.

This commit should address the issues, but this was done in a very
limited test. Package changes detection still works fine and there was
no log vomit from the system — which means no regression as far as I can
see.

For now, I'll keep this until there's more report coming in.