Stop using receivers altogether
I don't want to describe how messy and hacky this workaround is, but I
just can't handle the inconsistency of receivers. Due to how strict
Android has become with receivers, it's getting harder to properly
handle implicit receivers without going to JobSchedulers.

The hack here is to simply count the installed packages and compare that
with the number fetched during onCreate. This effectively works for
lifecycle switches (pauses, resumes, restarts) but not when an app is
installed while the launcher is running.

I plan on bringing receiver back, but it'll only be used to handle
in-foreground changes, which this hack can't do.