Fix hanging settings when changing a preference
The regression occured in Android P. If the user changes the theme
repeatedly and then proceeds to change another preference (say, a
switch), the switch will causes a soft lock-up, freezing the launcher
entirely.

Fix this by handling finishes manually. I don't know if this is the best
solution, but it works for now.