Make receiver work reliably with package counter
The receiver here is no longer used to listen to background changes.
Everything is done in the foreground. If there is an app installed when
the launcher is running, the receiver handles it, but if it happens in
the background, we can simply check it when we return to the foreground
using the package counter.

Additional changes here include removing the 'PACKAGE_REMOVED' intent
data. Removing a package will always trigger onPause and onResume, which
means the package counter will be at play here.

The receiver also now does the activity recreation itself. This should
make it more bulletproof as we also rely less on preferences here.