Refactor LauncherAppWidgetHostView to fix issues
The refactor here addresses all issues I can remember of, which mostly
consist of broken tap, long press, and swipe handling.

The issue stems from the limitation of using MotionEvent solely to check
for the event themselves. MotionEvent is only good for checking a
limited set of events, most of them collide with each other or will need
a (somewhat) of a huge workaround to work on.

Our previous solution (in a commit before this) was refactoring
Runnables, leaving the MotionEvent check intact.

This commit follows suit by refactoring the class to use GestureDetector
— a class which can handle plenty of events with a greater accuracy.
With this change, the code is also cleaner and much more succinct.