Don't recreate preference listener
Whenever the launcher (or rather, SettingsActivity) is restarted by a
theme change, MainActivity will follow suit as well. This causes
MainActivity to reinitialise a lot of its inner workings at theme
restart.

Previously I made a commit addressing the lock-up of Settings in
subsequent theme restarts. It was largely a hack as I didn't know what
precisely caused it.

It seems that after quite the trial and error, I found that
PreferenceChangeListener is the culprit here. Having a simple static
boolean that checks for its existence, does indeed, fixes the freezing
Settings — all without needing a hack.

I do sort of miss the slightly faster Settings. That may come at a later
commit if possible.