Don't call clearViews in stopListening
Some widgets, such as those originating from a music/media player can
have their state frozen in place (i.e a suspended play button) when the
launcher is put to the background.

If we call clearViews() here and the launcher is suspended, the view of
the widget will be stuck to whatever is shown before the launcher is put
to the background. So, let's not do that and fix the issue.