Migrate to preference-v7
We were using a rather old library to make consistent-looking
preferences for pre-Lollipop devices. The library had some derelict
function that severely limited the preference function, causing the need
for odd workarounds.

Switching over to Google's preference-v7 allows us to leverage extra
customisability, and additionally, we can finally remove those
workarounds.

Included in this commit is also a better workaround for the sluggish
settings caused by the constant recreate when changing theme. The
Listener was removed in favour of activity request code; this feels a
bit excessive, but it should do miles better than the messy Listener we
had.