Fix unapplied changes when SetingsActivity is called elsewhere
When SettingsActivity is called from outside the launcher, then the
changes it make will not directly apply the launcher, as it doesn't
trigger to onActivityResult we use to detect changes.

So instead, we need to track the caller of the activity — this is
because when called from the launcher, SettingsActivity will have a
calling package. Anywhere else (say, the system's Settings) won't.