Allow pinned items to be re-ordered
FlexibleAdapter supports reordering of items, we just haven't got around
to implementing it. This should allow the favourites panel items to be
reorganised by a long press and a drag.

Foreach may not be the smartest idea here, and I can see other ways of
updating it without getting the entire list, but this is fine for the
time being.

As requested by #6.