Unbreak disabled/enabled apps
When an app is disabled through settings,the system sends out
PACKAGE_CHANGED. This is fine, except that even when the receiver is set
out to receive this intent, it rarely ever receives it for
user-initiated action (e.g, the user goes to Settings and disable a
package).

A super quick workaround here is to simply iterate over all installed
packages in countInstalledPackage. By filtering disabled package(s), the
method should return a changed value when a package is disabled/changed
— triggering a list refresh.