Implement basic shorthand system
This implementation allows apps to be tagged using a specified label.
Apps can then be searched through the label instead of its name. This
hopefully should be able to save a few keyboard strokes for app search.

With this, it is also possible to group apps based off of a similar
shorthands. For instance, several browser apps can be shorthanded to
'browser', and a search for 'browser' will return them all.

Should somewhat satisfies the third point in #11, and may be enough for