Guard against stray '|' when setting shorthand
While this won't crash anything, it's much better if we strip out the
vbars because they may pose future issues down the line.

It may also be better if we warn the user about the invalid character,
as opposed to directly stripping them here. That may come in a future
commit.