Move widgets to its own fragment
It's somewhat implausible in the launcher's current state to host an
interactive widget while keeping the current navigation flow. This is
worsened by the ongoing attempt at implementing multiple widgets
support.

The commit here puts widgets in a separate space (a DialogFragment) to
allow for easier and more maintainable management. Because it is a
separate view altogether, we don't need to worry about breaking
navigation.

Upon testing, there has been no regression with this switch. The codes
may need cleaning, however.