Fix crashing when widgets are removed
When widgets are removed arbitrarily (not in their order), the launcher
will crash because of the different index presented by the list and the
widgets. Because of this, we need another unique index to make sure the
alteration of the list stay within the bounds of its size.

This commit should fix the issue presented above, introducing the
widget's ID as another unique index to present when removing widgets.