[EXP] Implement customisable web search provider
The implementation is fairly simple and it keeps most of our methods
from previous, fixed-value providers.

With this, the users can add their own search provider, not restricted
by the three defaults set in the launcher. The defaults will still be
available, but they are now removable and can be edited. Currently, it
is not possible to completely empty out the whole list due to risk of
crashing the launcher.

The providers are currently sorted the way HashSet feels like sorting
it. It's mostly a minor cosmetic issue, and proper list sorting is not
in mind for this commit.

There may need to be lots of methods cleanup in PreferenceHelper, as I
see that it has become bloated with similar-looking functions. These are
to be addressed at later date.