Use own Snackbar class
This eliminates the need for the hackish disableSnackbarSwipe method.
Although it's arguably not at all less hackish, it should work more
reliably.

The DagashiBar code is ripped straight from the support library. This
isn't an optimal way of implementing it, but there's no other way as
Snackbar is a final class in the support library.