Implement swipe to expand status bar
There isn't a single official API to do this. The usual method
(reflection) is employed here. This should work from our min SDK (ICS)
all the way to the most recent public SDK (Pie, in time of this
writing).

Because of its activation conflicting with the usual swipe-down gesture,
this requires tap-to-open to be enabled — as it will confuse users if
they are locked out of their app drawer.