Improve custom web provider support
Some search providers URL may have its query not placed directly at the
end (for instance: test.url/search?q={query}&other=things). This format
breaks custom provider support, and it shows the inflexibility of the
system we have.

Fix this by having proper query placeholder in place. With this, the
user only need to make sure the URL has a placeholder (%s), and the
launcher will automatically replace the placeholder with the query.