Prevent searchSnack from closing early
When the user taps/touches a suggestion from their keyboard, it is
almost certain that searchSnack will close. This is because selecting a
suggestion empties the bar, which counts as a reason to dismiss
searchSnack.

Fix this by using a timer to keep track of user's inactivity. This
should avoid the issue for now, as the suggestion done by keyboard would
complete faster than the delay time.