Make shortcut ID unique
It is possible (though implausible) that an app has a shortcut with the
same label. If this were to occur, shortcutMap will overwrite its
previous value. This may cause a crash.

To fix this, assign an incrementing ID starting from
SHORTCUT_MENU_GROUP's value. This should guarantee a sense of uniqueness
without overcomplicating the code.