Extend gesture actions
This commit further extends available gesture actions for the launcher.
On the launcher's side, this adds an option to open Widgetspace, and
further options can be implemented by a separate app ('gesture handler')
— this allows the launcher to be free of the burden of requesting
plentiful permissions, as well as making the launcher's code to be more
manageable.

Gesture handlers will be documented in a future commit, and perhaps in a
separate repository.