Organise preferences to subscreens
This has been long overdue now that we have a heaping pile of
preferences in store.

This commit needed a hack in PreferenceFragment for it to work. The
abundance of null-checks there are about as temporary as I think they
are since we can check for rootKey instead of relying on findPreference
nullity — that may be handled later in a future commit.